+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.1.4](https://github.com/bitrix24/b24ui/compare/v2.1.3...v2.1.4) (2025-11-xx)
 
 * fix(module): scan layers when using component detection
+* fix(ColorModeButton): use css to display color mode icon
 
 ## [2.1.3](https://github.com/bitrix24/b24ui/compare/v2.1.2...v2.1.3) (2025-11-17)