+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.1.5](https://github.com/bitrix24/b24ui/compare/v2.1.4...v2.1.5) (2025-11-xx)
+
+* docs(navigation-menu): incorrect index in model value example
+
 ## [2.1.4](https://github.com/bitrix24/b24ui/compare/v2.1.3...v2.1.4) (2025-11-18)
 
 ### Features